[安全]后台模板管理编辑有越过二次密保权限的问题； [新增]支持[商品采集]插件； [新增]支持[短信通知]插件； [新增]支持[站群快捷管理]插件； [新增]支持[后台定时改密提醒]插件； [新增]抖音小程序订单同步问题； [优化]广告位标签的pid属性支持变量； [优化]小程序订单处理异常问题； [优化]外贸助手新增（图）的语言变量； [优化]前台自定义字段标签显示的兼容问题； [优化]前台seo分页标题支持内置的外贸助手语言变量； [优化]原生小程序支持抖音支付； [优化]多城市分站支持当前区域可以显示所有下级区域的文档； [优化]开发模式下，不使用大量数据缓存，减少runtime/cache目录占用空间； [优化]后台内容管理时，发布产品模型的文档弹窗选择框，支持永久关闭不再提示； [优化]后台订单详情新增展示物流单号； [优化]后台云插件库在网络异常时也显示； [优化]后台增加自动处理订单逻辑； [优化]后台栏目新建模板与安全验证密码的冲突体验； [优化]后台发布/编辑文档时，自动提取seo描述不再替换空格； [优化]后台发布、编辑文档的自定义文件名由全站判断改成同栏目下，不能出现同名自定义文件名； [优化]支付、售后的相关功能，并售后支持可视化商城小程序； [优化]支付后订单的短信、邮箱提醒； [优化]商品规格图片； [优化]多城市站点一键开启或关闭二级域名； [优化]兼容微站点和微信公众号扫码登录(插件)； [优化]公众号登录不能支付/公众号登录和微站点登录不是一个账号； [修复]在运营模式下微信、支付宝支付的异步回调失效； [修复]前台留言表单字段名称在单引号情况下报错； [修复]在伪静态下，前台多语言切换到非中文语言下，刷新时报404，时正常； [修复]微信支付后台配置验证问题； [修复]多城市分站的面包屑链接错误； [修复]多城市在会员中心收藏列表里的文档链接错误； [修复]多城市一些已知BUG； [修复]小程序修改手机号码问题； [修复]开启绝对路径下，前台内容的远程图片显示不了； [修复]子目录网站的前后台详情页图片不显示； [修复]后台功能地图添加主题风格、外贸助手到左侧菜单的bug； [修复]后台订单导出时，有些网站报错； [修复]后台删除留言栏目时，表单数据也一起被误删了； [修复]后台API配置微信小程序总报“支付商户号或支付密钥不正确！” [修复]后台多语言情况下，在其他语言修改旧版产品参数的排序号不生效； [修复]后台入口文件更名后，手工生成网站地图报404/首页信息； [修复]后台自定义模型没有添加html类型字段时，多图懒加载不生效； [修复]后台勾选更新上下篇后，发布/编辑文档生成报错； [修复]后台广告管理禁用失效； [修复]后台测试发送邮箱、短信报错；

+++ application/admin/logic/WeappLogic.php
@@ -23,10 +23,11 @@ class WeappLogic extends Model
     public $data_path;
     // public $config_path;
     // public $curent_version;    
-    public $service_url;
-    // public $upgrade_url;
+    public $upgrade_url;
     public $service_ey;
     // public $code;
+    public $cms_version;
+    public $upgrade_postdata = [];
     
     /**
      * 析构函数
@@ -34,15 +35,20 @@ class WeappLogic extends Model
     function  __construct() {
         // $this->code = input('param.code/s', '');
         // $this->curent_version = getWeappVersion($this->code);
-        $this->service_ey = config('service_ey');
+        $this->cms_version = getCmsVersion();
+        $upgradeLogic = new \app\admin\logic\UpgradeLogic;
+        $this->service_ey = $upgradeLogic->getServiceUrl(true);
         $this->root_path = ROOT_PATH; // 
         $this->weapp_path = WEAPP_DIR_NAME.DS; // 
         $this->data_path = DATA_PATH; // 
         // $this->config_path = $this->weapp_path.$this->code.DS.'config.php'; // 版本配置文件路径
         // api_Weapp_checkVersion
-        $tmp_str = 'L2luZGV4LnBocD9tPWFwaSZjPVdlYXBwJmE9Y2hlY2tWZXJzaW9u';
-        $this->service_url = base64_decode($this->service_ey).base64_decode($tmp_str).'&domain='.request()->host(true);
-        // $this->upgrade_url = $this->service_url.'&code='.$this->code.'&v='.$this->curent_version;
+        $this->upgrade_url = $this->service_ey.'/index.php?m=api&c=Weapp&a=checkVersion';
+        $this->upgrade_postdata = [
+            'domain' => request()->host(true),
+            'cms_version' => $this->cms_version,
+            'ip' => serverIP(),
+        ];
     }
 
     /**
@@ -51,7 +57,7 @@ class WeappLogic extends Model
      */
     public function insertWeapp()
     {
-        $row        = Db::name('weapp')->field('id,code,config,is_buy')->getAllWithIndex('code'); // 数据库
+        $row        = Db::name('weapp')->field('*')->getAllWithIndex('code'); // 数据库
         $new_arr    = array(); // 本地
         $addData    = array(); // 数据存储变量
         $updateData = array(); // 数据存储变量
@@ -92,7 +98,7 @@ class WeappLogic extends Model
         }
         //数据库有 本地没有
         foreach($row as $k => $v){
-            if (!in_array($v['code'], $new_arr) && $v['is_buy'] < 1) {//is_buy  0->本地安装,1-线上购买
+            if (!in_array($v['code'], $new_arr) && isset($v['is_buy']) && $v['is_buy'] < 1) {//is_buy  0->本地安装,1-线上购买
                 Db::name('weapp')->where($v)->cache(true, null, 'weapp')->delete();
             }
         }
@@ -180,13 +186,29 @@ class WeappLogic extends Model
     public function checkVersion($code, $serviceVersionList = false) {
         error_reporting(0);//关闭所有错误报告
         $lastupgrade = array();
+
         if (false === $serviceVersionList) {
-            $curent_version = getWeappVersion($code);
-            $url = $this->service_url.'&code='.$code.'&v='.$curent_version;
-            $context = stream_context_set_default(array('http' => array('timeout' => 3,'method'=>'GET')));
-            $serviceVersionList = @file_get_contents($url,false,$context);    
-            $serviceVersionList = json_decode($serviceVersionList,true);
+            $this->upgrade_postdata['code'] = $code;
+            $this->upgrade_postdata['v'] = getWeappVersion($code);
+            $upgradeLogic = new \app\admin\logic\UpgradeLogic;
+            $upgradeLogic->GetKeyData($this->upgrade_postdata);
+            $serviceVersionList = @httpRequest($this->upgrade_url, 'POST', $this->upgrade_postdata, [], 5);
+            if (false === $serviceVersionList) {
+                $url = $this->upgrade_url.'&'.http_build_query($this->upgrade_postdata);
+                $context = stream_context_set_default(array('http' => array('timeout' => 5,'method'=>'GET')));
+                $serviceVersionList = @file_get_contents($url, false, $context);
+            }
+            if (false === $serviceVersionList) {
+                return ['code' => 0, 'msg' => "无法连接远程升级服务器！"];
+            } else {
+                $serviceVersionList = json_decode($serviceVersionList,true);
+                if (isset($serviceVersionList['code']) && empty($serviceVersionList['code'])) {
+                    $msg = empty($serviceVersionList['msg']) ? 'API请求超时' : $serviceVersionList['msg'];
+                    return ['code' => 0, 'msg' => "<font color='red'>{$msg}</font>"];
+                } 
+            }
         }
+
         if(!empty($serviceVersionList))
         {
             $upgradeArr = array();
@@ -264,14 +286,31 @@ class WeappLogic extends Model
         }
 
         $curent_version = getWeappVersion($code);
-        $upgrade_dev = config('global.upgrade_dev');
-        $upgrade_url = $this->service_url.'&code='.$code.'&dev='.$upgrade_dev.'&v='.$curent_version;
-        $serviceVersionList = file_get_contents($upgrade_url);
-        $serviceVersionList = json_decode($serviceVersionList,true);
-        if (empty($serviceVersionList)) {
-            return ['code' => 0, 'msg' => "没找到升级包"];
+        $this->upgrade_postdata['code'] = $code;
+        $this->upgrade_postdata['dev'] = config('global.upgrade_dev');
+        $this->upgrade_postdata['v'] = $curent_version;
+        $upgradeLogic = new \app\admin\logic\UpgradeLogic;
+        $upgradeLogic->GetKeyData($this->upgrade_postdata);
+        $url = $this->service_ey."/index.php?m=api&c=Weapp&a=upgradeVersion";
+        $serviceVersionList = @httpRequest($url, 'POST', $this->upgrade_postdata, [], 5);
+        if (false === $serviceVersionList) {
+            $url = $url.'&'.http_build_query($this->upgrade_postdata);
+            $context = stream_context_set_default(array('http' => array('timeout' => 5,'method'=>'GET')));
+            $serviceVersionList = @file_get_contents($url, false, $context);
         }
-        
+        if (false === $serviceVersionList) {
+            return ['code' => 0, 'msg' => "无法连接远程升级服务器！"];
+        } else {
+            $serviceVersionList = json_decode($serviceVersionList,true);
+            if (isset($serviceVersionList['code']) && empty($serviceVersionList['code'])) {
+                $msg = empty($serviceVersionList['msg']) ? 'API请求超时' : $serviceVersionList['msg'];
+                return ['code' => 0, 'msg' => $msg];
+            }
+        }
+        if (empty($serviceVersionList)) {
+            return ['code' => 0, 'msg' => "当前没有可升级的版本！"];
+        }
+
         clearstatcache(); // 清除文件夹权限缓存
         $config_path = $this->weapp_path.$code.DS.'config.php'; // 版本配置文件路径
         if(!is_writeable($config_path)) {
@@ -328,7 +367,7 @@ class WeappLogic extends Model
 
             /*更新的SQL文件列表*/
             $sql_file = !empty($val['sql_file']) ? $val['sql_file'] : array();
-            $sqlfileArr = array_merge($sqlfileArr, $val['sql_file']);
+            $sqlfileArr = array_merge($sqlfileArr, $sql_file);
             /*--end*/
         }
         /*--end*/
@@ -513,27 +552,8 @@ class WeappLogic extends Model
      */
     private function downloadFile($fileUrl,$md5File)
     {
-        $downFileName = explode('/', $fileUrl);    
-        $downFileName = end($downFileName);
-        $saveDir = $this->data_path.'backup'.DS.$downFileName; // 保存目录
-        tp_mkdir(dirname($saveDir));
-        if(!file_get_contents($fileUrl, 0, null, 0, 1)){
-            return ['code' => 0, 'msg' => '官方插件升级包不存在']; // 文件存在直接退出
-        }
-        $ch = curl_init($fileUrl);            
-        curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-        curl_setopt($ch, CURLOPT_BINARYTRANSFER,1);
-        $file = curl_exec ($ch);
-        curl_close ($ch);                                                            
-        $fp = fopen($saveDir,'w');
-        fwrite($fp, $file);
-        fclose($fp);
-        if(!eyPreventShell($saveDir) || !file_exists($saveDir) || $md5File != md5_file($saveDir))
-        {
-              // 下载保存升级包失败，请检查所有目录的权限以及用户组不能为root
-            return ['code' => 0, 'msg' => 'data目录不可写入，无法执行更新，请检查该目录权限以及用户组再重试'];
-        }
-        return ['code' => 1, 'msg' => '下载成功'];
+        $upgradeLogic = new \app\admin\logic\UpgradeLogic;
+        return $upgradeLogic->downloadFile($fileUrl, $md5File, 'weapp');
     }            
     
     // 升级记录 log 日志
@@ -547,8 +567,8 @@ class WeappLogic extends Model
         }
         $mysqlinfo = \think\Db::query("SELECT VERSION() as version");
         $mysql_version  = $mysqlinfo[0]['version'];
-        $vaules = array(
-            'domain'=>$_SERVER['HTTP_HOST'], //用户域名  
+        $values = array(
+            'domain'=>request()->host(), //用户域名  
             'code'  => $code, // 插件标识
             'key_num'=>getWeappVersion($code), // 用户版本号
             'to_key_num'=>$to_key_num, // 用户要升级的版本号                
@@ -560,9 +580,114 @@ class WeappLogic extends Model
             'web_server'    => $_SERVER['SERVER_SOFTWARE'],
         );
         // api_Weapp_upgradeLog
-        $tmp_str = 'L2luZGV4LnBocD9tPWFwaSZjPVdlYXBwJmE9dXBncmFkZUxvZyY=';
-        $url = base64_decode($this->service_ey).base64_decode($tmp_str).http_build_query($vaules);
-        file_get_contents($url);
+        $upgradeLogic = new \app\admin\logic\UpgradeLogic;
+        $upgradeLogic->GetKeyData($values);
+        $url = $this->service_ey.'/index.php?m=api&c=Weapp&a=upgradeLog';
+        httpRequest($url, 'POST', $values, [], 5);
+    }
+
+    /**
+     * 兼容早期版本，1.4.5或以下版本
+     * @param  array  &$assign_data [description]
+     * @return [type]               [description]
+     */
+    public function installpwd_145(&$assign_data = [])
+    {
+        $main_lang = get_main_lang();
+        /*插件安装密码设置*/
+        $weapp_installpwd = tpCache('weapp.weapp_installpwd', [], $main_lang);
+        if (!empty($weapp_installpwd)) {
+            $weapp_installpwd = 1;
+        } else {
+            $weapp_installpwd = 0;
+        }
+        $assign_data['weapp_installpwd'] = $weapp_installpwd;
+        /*--end*/
+
+        $admin_info = session('admin_info');
+        /*是否创始人*/
+        $isFounder = 0;
+        if (empty($admin_info['parent_id']) && -1 == $admin_info['role_id']) {
+            $isFounder = 1;
+        }
+        $assign_data['isFounder'] = $isFounder;
+        /*--end*/
+
+        /*登录第一次输入插件安装密码之后，在退出之前安装所有插件都不再输入安装密码*/
+        $is_weapp_installpwd = 0;
+        $weapp_installpwd = tpCache('weapp.weapp_installpwd', [], $main_lang);
+        $firstInstallpwd = empty($admin_info['weapp_info']['firstInstallpwd']) ? '' : $admin_info['weapp_info']['firstInstallpwd'];
+        if (!empty($firstInstallpwd) && $firstInstallpwd == $weapp_installpwd) {
+            $is_weapp_installpwd = 1;
+        }
+        $assign_data['is_weapp_installpwd'] = $is_weapp_installpwd;
+        /*--end*/
+    }
+
+    /**
+     *  加密函数
+     *
+     * @access    public
+     * @param     string $string 字符串
+     * @param     string $operation 操作
+     * @return    string
+     */
+    public function mchStrCode($string, $operation = 'ENCODE', $auth_code = '')
+    {
+        if (empty($auth_code)) {
+            if (function_exists('get_auth_code')) {
+                $auth_code = get_auth_code();
+            } else {
+                $auth_code = tpCache('system.system_auth_code');
+                if (empty($auth_code)) {
+                    $auth_code = \think\Config::get('AUTH_CODE');
+                    /*多语言*/
+                    if (is_language()) {
+                        $langRow = \think\Db::name('language')->order('id asc')->select();
+                        foreach ($langRow as $key => $val) {
+                            tpCache('system', ['system_auth_code'=>$auth_code], $val['mark']);
+                        }
+                    } else { // 单语言
+                        tpCache('system', ['system_auth_code'=>$auth_code]);
+                    }
+                    /*--end*/
+                }
+            }
+        }
+        
+        if (function_exists('mchStrCode')) {
+            return mchStrCode($string, $operation, $auth_code);
+        }
+
+        $key_length = 4;
+        $expiry     = 0;
+        $key        = md5($auth_code);
+        $fixedkey   = md5($key);
+        $egiskeys   = md5(substr($fixedkey, 16, 16));
+        $runtokey   = $key_length ? ($operation == 'ENCODE' ? substr(md5(microtime(true)), -$key_length) : substr($string, 0, $key_length)) : '';
+        $keys       = md5(substr($runtokey, 0, 16) . substr($fixedkey, 0, 16) . substr($runtokey, 16) . substr($fixedkey, 16));
+        $string     = $operation == 'ENCODE' ? sprintf('%010d', $expiry ? $expiry + time() : 0) . substr(md5($string . $egiskeys), 0, 16) . $string : base64_decode(substr($string, $key_length));
+
+        $i             = 0;
+        $result        = '';
+        $string_length = strlen($string);
+        for ($i = 0; $i < $string_length; $i++) {
+            $result .= chr(ord($string[$i]) ^ ord($keys[$i % 32]));
+        }
+        if ($operation == 'ENCODE') {
+            return $runtokey . str_replace('=', '', base64_encode($result));
+        } else {
+            $str1 = substr($result, 0, 10);
+            if(version_compare(PHP_VERSION,'8.0.0','>')) {
+                $str1 = !empty($str1) ? $str1 : 0;
+            }
+
+            if (($str1 == 0 || intval($str1) - time() > 0) && substr($result, 10, 16) == substr(md5(substr($result, 26) . $egiskeys), 0, 16)) {
+                return substr($result, 26);
+            } else {
+                return '';
+            }
+        }
     }
 } 
 ?>
